+++ app/interactors/check_removed_head_role.rb
@@ -7,8 +7,8 @@ class CheckRemovedHeadRole
   def call
     return if context.guild_role.name != 'gm'
     user = context.guild_role.character.user
-    other_head_characters_of_user_in_guild = user.characters.where.not(id: context.guild_role.character_id).joins(:guild_role).where(guild_roles: { name: 'gm' })
-    return unless other_head_characters_of_user_in_guild.empty?
+    other_head_characters_of_user = user.characters.where.not(id: context.guild_role.character_id).joins(:guild_role).where(guild_roles: { name: 'gm' })
+    return unless other_head_characters_of_user.empty?
     notification = Notification.find_by(event: 'guild_request_creation', status: 1)
     return if notification.nil?
     Delivery.where(deliveriable: user, notification: notification).destroy_all